Refactor lava lamp into its own file and call from @application.cpp instead fixes #11
