+++ work/src/application.cpp
@@ -42,558 +42,13 @@ Application::Application(GLFWwindow* window) : m_window(window) {
 	//m_model.mesh = load_wavefront_data(CGRA_SRCDIR + std::string("/res//assets//teapot.obj")).build();
 	m_model.color = vec3(1, 0, 0);
 
-	// Initialize lava lamp
-	initializeLavaLamp();
+	// Initialize lava lamp using new LavaLamp API
+	m_lavaLamp.initialiseLavaLamp(
+		CGRA_SRCDIR + std::string("//res//shaders//lava_vertex.glsl"),
+		CGRA_SRCDIR + std::string("//res//shaders//lava_fragment.glsl")
+	);
 }
 
-void Application::ensureDepthFBO(int width, int height) {
-	// If existing size matches, nothing to do
-	if (m_depthFBO != 0 && m_depthTexW == width && m_depthTexH == height)
-		return;
-
-	// Delete old if present
-	if (m_depthFBO != 0) {
-		glDeleteFramebuffers(1, &m_depthFBO);
-		m_depthFBO = 0;
-	}
-	if (m_depthTextureFront != 0) {
-		glDeleteTextures(1, &m_depthTextureFront);
-		m_depthTextureFront = 0;
-	}
-	if (m_depthTextureBack != 0) {
-		glDeleteTextures(1, &m_depthTextureBack);
-		m_depthTextureBack = 0;
-	}
-
-	// Create two depth textures (front and back)
-	glGenTextures(1, &m_depthTextureFront);
-	glBindTexture(GL_TEXTURE_2D, m_depthTextureFront);
-	glTexImage2D(GL_TEXTURE_2D, 0, GL_DEPTH_COMPONENT24, width, height, 0, GL_DEPTH_COMPONENT, GL_UNSIGNED_INT, nullptr);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
-	// Ensure sampler sampling returns float not depth-compare on some drivers
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_COMPARE_MODE, GL_NONE);
-	// define base/max level to avoid "no base level" warnings
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_BASE_LEVEL, 0);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAX_LEVEL, 0);
-
-	glGenTextures(1, &m_depthTextureBack);
-	glBindTexture(GL_TEXTURE_2D, m_depthTextureBack);
-	glTexImage2D(GL_TEXTURE_2D, 0, GL_DEPTH_COMPONENT24, width, height, 0, GL_DEPTH_COMPONENT, GL_UNSIGNED_INT, nullptr);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_COMPARE_MODE, GL_NONE);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_BASE_LEVEL, 0);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAX_LEVEL, 0);
-
-	glBindTexture(GL_TEXTURE_2D, 0);
-
-	// Create FBO (we will attach either texture as depth attachment for each pass)
-	glGenFramebuffers(1, &m_depthFBO);
-	m_depthTexW = width;
-	m_depthTexH = height;
-}
-
-void Application::initializeLavaLamp() {
-	// Build lava lamp shader
-	shader_builder lava_sb;
-	lava_sb.set_shader(GL_VERTEX_SHADER, CGRA_SRCDIR + std::string("//res//shaders//lava_vertex.glsl"));
-	lava_sb.set_shader(GL_FRAGMENT_SHADER, CGRA_SRCDIR + std::string("//res//shaders//lava_fragment.glsl"));
-	m_lavaShader = lava_sb.build();
-
-	// Initialize the lava lamp simulation with 5 blobs
-	m_lavaLamp.initialize(5);
-
-	m_threshold = 1.0f;         // Lower threshold
-	m_heaterTemp = 120.0f;      // increased heater for stronger rise
-	m_gravity = -9.8f;          // fixed gravity (permanent)
-
-	// Apply these settings to the lamp (set once)
-	m_lavaLamp.setThreshold(m_threshold);
-	m_lavaLamp.setHeaterTemperature(m_heaterTemp);
-	m_lavaLamp.setGravity(m_gravity);
-	// viscosity is automatic now; no setViscosity call
-
-	// Create lamp container mesh (wrap meshes in basic_model so draw() sets shader + matrices)
-	m_lampGlassModel.mesh = createLampContainerGlass();
-	m_lampGlassModel.shader = m_lavaShader;
-	m_lampGlassModel.modelTransform = glm::mat4(1.0f);
-
-	m_lampMetalModel.mesh = createLampContainerMetal();
-	m_lampMetalModel.shader = m_lavaShader;
-	m_lampMetalModel.modelTransform = glm::mat4(1.0f);
-
-	m_fullscreenQuadModel.mesh = createFullscreenQuad();
-	m_fullscreenQuadModel.shader = m_lavaShader;
-	m_fullscreenQuadModel.modelTransform = glm::mat4(1.0f);
-
-
-	// Depth FBO will be created lazily in render (size depends on framebuffer)
-	m_depthFBO = 0;
-	m_depthTextureFront = 0;
-	m_depthTextureBack = 0;
-	m_depthTexW = m_depthTexH = 0;
-
-	// Set initial time
-	m_lastTime = static_cast<float>(glfwGetTime());
-}
-
-cgra::gl_mesh Application::createFullscreenQuad() {
-	mesh_builder builder;
-
-	// Create a large quad that covers the screen
-	mesh_vertex v0, v1, v2, v3;
-	v0.pos = vec3(-1, -1, 0); v0.uv = vec2(0, 0);
-	v1.pos = vec3(1, -1, 0);  v1.uv = vec2(1, 0);
-	v2.pos = vec3(1, 1, 0);   v2.uv = vec2(1, 1);
-	v3.pos = vec3(-1, 1, 0);  v3.uv = vec2(0, 1);
-
-	builder.push_vertex(v0);
-	builder.push_vertex(v1);
-	builder.push_vertex(v2);
-	builder.push_vertex(v3);
-
-	builder.push_index(0); builder.push_index(1); builder.push_index(2);
-	builder.push_index(0); builder.push_index(2); builder.push_index(3);
-
-	return builder.build();
-}
-
-cgra::gl_mesh Application::createLampContainerGlass() {
-	mesh_builder builder;
-	const int segments = 64;
-
-	struct ProfilePoint {
-		float height;
-		float radius;
-	};
-
-	// Only bottom and top rings (smooth glass shell)
-	std::vector<ProfilePoint> profile = {
-		{1.7f, 1.8f},  // bottom of glass (just above the metal section)
-		{10.0f, 1.0f}  // top ring
-	};
-
-	std::vector<int> ringStarts;
-	for (size_t p = 0; p < profile.size(); ++p) {
-		ringStarts.push_back(builder.vertices.size());
-		for (int i = 0; i <= segments; ++i) {
-			float angle = 2.0f * pi<float>() * i / segments;
-			float x = profile[p].radius * cos(angle);
-			float z = profile[p].radius * sin(angle);
-			float y = profile[p].height;
-
-			mesh_vertex v;
-			v.pos = vec3(x, y, z);
-			v.norm = normalize(vec3(x, 0.0f, z));
-			v.uv = vec2(float(i) / segments, float(p) / (profile.size() - 1));
-			builder.push_vertex(v);
-		}
-	}
-
-	// Connect the two rings into quads
-	int currentRing = ringStarts[0];
-	int nextRing = ringStarts[1];
-	for (int i = 0; i < segments; ++i) {
-		int curr = currentRing + i;
-		int next = currentRing + i + 1;
-		int currAbove = nextRing + i;
-		int nextAbove = nextRing + i + 1;
-
-		builder.push_index(curr);
-		builder.push_index(currAbove);
-		builder.push_index(next);
-
-		builder.push_index(next);
-		builder.push_index(currAbove);
-		builder.push_index(nextAbove);
-	}
-
-	return builder.build();
-}
-
-cgra::gl_mesh Application::createLampContainerMetal() {
-	mesh_builder builder;
-	const int segments = 64;
-
-	// Lower third of bulb (metal)
-	{
-		struct ProfilePoint {
-			float height;
-			float radius;
-		};
-
-		std::vector<ProfilePoint> lowerBulb = {
-			{0.0f, 1.2f},  // bottom ring
-			{1.7f, 1.8f}   // cut height
-		};
-
-		std::vector<int> ringStarts;
-		for (size_t p = 0; p < lowerBulb.size(); ++p) {
-			ringStarts.push_back(builder.vertices.size());
-			for (int i = 0; i <= segments; ++i) {
-				float angle = 2.0f * pi<float>() * i / segments;
-				float x = lowerBulb[p].radius * cos(angle);
-				float z = lowerBulb[p].radius * sin(angle);
-				float y = lowerBulb[p].height;
-
-				mesh_vertex v;
-				v.pos = vec3(x, y, z);
-				v.norm = normalize(vec3(x, 0.0f, z));
-				v.uv = vec2(float(i) / segments, float(p) / (lowerBulb.size() - 1));
-				builder.push_vertex(v);
-			}
-		}
-
-		for (size_t p = 0; p < lowerBulb.size() - 1; ++p) {
-			int currentRing = ringStarts[p];
-			int nextRing = ringStarts[p + 1];
-			for (int i = 0; i < segments; ++i) {
-				int curr = currentRing + i;
-				int next = currentRing + i + 1;
-				int currAbove = nextRing + i;
-				int nextAbove = nextRing + i + 1;
-
-				builder.push_index(curr);
-				builder.push_index(currAbove);
-				builder.push_index(next);
-
-				builder.push_index(next);
-				builder.push_index(currAbove);
-				builder.push_index(nextAbove);
-			}
-		}
-	}
-
-	// Glass bottom cap (inverted)
-	{
-		float capHeight = 0.0f;
-		float capDepth = 0.8f;
-		float capRadius = 1.2f;
-		int capStart = builder.vertices.size();
-
-		for (int i = 0; i <= segments; ++i) {
-			float angle = 2.0f * pi<float>() * i / segments;
-			float xOuter = capRadius * cos(angle);
-			float zOuter = capRadius * sin(angle);
-
-			mesh_vertex vTop;
-			vTop.pos = vec3(xOuter, capHeight, zOuter);
-			vTop.norm = normalize(vec3(xOuter, 0, zOuter));
-			vTop.uv = vec2(float(i) / segments, 0.0f);
-
-			mesh_vertex vBottom;
-			vBottom.pos = vec3(0, capHeight - capDepth, 0);
-			vBottom.norm = vec3(0, -1, 0);
-			vBottom.uv = vec2(0.5f, 1.0f);
-
-			builder.push_vertex(vTop);
-			builder.push_vertex(vBottom);
-		}
-
-		for (int i = 0; i < segments; ++i) {
-			int i0 = capStart + i * 2;
-			int i1 = capStart + i * 2 + 1;
-			int i2 = capStart + (i + 1) * 2;
-			int i3 = capStart + (i + 1) * 2 + 1;
-
-			builder.push_index(i0);
-			builder.push_index(i2);
-			builder.push_index(i1);
-
-			builder.push_index(i1);
-			builder.push_index(i2);
-			builder.push_index(i3);
-		}
-	}
-
-	// Glass top cap
-	{
-		float capBottomY = 10.0f;
-		float capTopY = 11.0f;
-		float capRadiusBottom = 1.0f;
-		float capRadiusTop = 0.8f;
-		int capStart = builder.vertices.size();
-
-		for (int i = 0; i <= segments; ++i) {
-			float angle = 2.0f * pi<float>() * i / segments;
-			float x1 = capRadiusBottom * cos(angle);
-			float z1 = capRadiusBottom * sin(angle);
-			float x2 = capRadiusTop * cos(angle);
-			float z2 = capRadiusTop * sin(angle);
-
-			mesh_vertex v1, v2;
-			v1.pos = vec3(x1, capBottomY, z1);
-			v2.pos = vec3(x2, capTopY, z2);
-			v1.norm = normalize(vec3(x1, 0, z1));
-			v2.norm = normalize(vec3(x2, 0, z2));
-
-			builder.push_vertex(v1);
-			builder.push_vertex(v2);
-		}
-
-		for (int i = 0; i < segments; ++i) {
-			int i0 = capStart + i * 2;
-			int i1 = capStart + i * 2 + 1;
-			int i2 = capStart + (i + 1) * 2;
-			int i3 = capStart + (i + 1) * 2 + 1;
-
-			builder.push_index(i0);
-			builder.push_index(i2);
-			builder.push_index(i1);
-
-			builder.push_index(i1);
-			builder.push_index(i2);
-			builder.push_index(i3);
-		}
-
-		int topCenter = builder.vertices.size();
-		mesh_vertex vTopCenter;
-		vTopCenter.pos = vec3(0, capTopY, 0);
-		vTopCenter.norm = vec3(0, 1, 0);
-		vTopCenter.uv = vec2(0.5f, 0.5f);
-		builder.push_vertex(vTopCenter);
-
-		for (int i = 0; i < segments; ++i) {
-			int outer0 = capStart + i * 2 + 1;
-			int outer1 = capStart + (i + 1) * 2 + 1;
-			builder.push_index(topCenter);
-			builder.push_index(outer0);
-			builder.push_index(outer1);
-		}
-	}
-
-	// Metal base
-	{
-		float baseHeight = -1.5f;
-		float baseRadiusBottom = 2.5f;
-		float baseRadiusTop = 1.2f;
-		int baseStart = builder.vertices.size();
-
-		for (int i = 0; i <= segments; ++i) {
-			float angle = 2.0f * pi<float>() * i / segments;
-			float x1 = baseRadiusBottom * cos(angle);
-			float z1 = baseRadiusBottom * sin(angle);
-			float y1 = baseHeight;
-
-			float x2 = baseRadiusTop * cos(angle);
-			float z2 = baseRadiusTop * sin(angle);
-			float y2 = 0.0f;
-
-			mesh_vertex v1, v2;
-			v1.pos = vec3(x1, y1, z1);
-			v2.pos = vec3(x2, y2, z2);
-			v1.norm = normalize(vec3(x1, 0, z1));
-			v2.norm = normalize(vec3(x2, 0, z2));
-			v1.uv = vec2(float(i) / segments, 0.0f);
-			v2.uv = vec2(float(i) / segments, 1.0f);
-
-			builder.push_vertex(v1);
-			builder.push_vertex(v2);
-		}
-
-		for (int i = 0; i < segments; ++i) {
-			int i0 = baseStart + i * 2;
-			int i1 = baseStart + i * 2 + 1;
-			int i2 = baseStart + (i + 1) * 2;
-			int i3 = baseStart + (i + 1) * 2 + 1;
-
-			builder.push_index(i0);
-			builder.push_index(i2);
-			builder.push_index(i1);
-
-			builder.push_index(i1);
-			builder.push_index(i2);
-			builder.push_index(i3);
-		}
-
-		int baseBottomCenter = builder.vertices.size();
-		mesh_vertex baseBottomV;
-		baseBottomV.pos = vec3(0, baseHeight, 0);
-		baseBottomV.norm = vec3(0, -1, 0);
-		baseBottomV.uv = vec2(0.5f, 0.5f);
-		builder.push_vertex(baseBottomV);
-
-		for (int i = 0; i < segments; ++i) {
-			int outer0 = baseStart + i * 2;
-			int outer1 = baseStart + (i + 1) * 2;
-			builder.push_index(baseBottomCenter);
-			builder.push_index(outer1);
-			builder.push_index(outer0);
-		}
-	}
-
-	return builder.build();
-}
-
-void Application::renderLavaLamp(const glm::mat4& view, const glm::mat4& proj) {
-
-
-	if (!m_showLavaLamp) return;
-
-
-	// Save current state
-	GLboolean depthMask;
-	glGetBooleanv(GL_DEPTH_WRITEMASK, &depthMask);
-	GLint depthFunc;
-	glGetIntegerv(GL_DEPTH_FUNC, &depthFunc);
-	GLboolean blendEnabled = glIsEnabled(GL_BLEND);
-
-
-	// retrieve the window size
-	int width, height;
-	glfwGetFramebufferSize(m_window, &width, &height);
-
-	// Ensure depth FBO/textures exist at current size
-	ensureDepthFBO(width, height);
-
-	// Update simulation
-	if (m_animateLamp) {
-		float currentTime = static_cast<float>(glfwGetTime());
-		float deltaTime = currentTime - m_lastTime;
-		m_lastTime = currentTime;
-		deltaTime = std::min(deltaTime, 0.05f);
-		m_lavaLamp.update(deltaTime);
-	}
-
-	glUseProgram(m_lavaShader);
-
-	// Set up matrices
-	mat4 model = mat4(1.0f);
-	mat4 modelView = view * model;
-	mat4 normalMatrix = transpose(inverse(model));
-
-	glUniformMatrix4fv(glGetUniformLocation(m_lavaShader, "uProjectionMatrix"), 1, GL_FALSE, value_ptr(proj));
-	glUniformMatrix4fv(glGetUniformLocation(m_lavaShader, "uModelViewMatrix"), 1, GL_FALSE, value_ptr(modelView));
-	glUniformMatrix4fv(glGetUniformLocation(m_lavaShader, "uModelMatrix"), 1, GL_FALSE, value_ptr(model));
-	glUniformMatrix4fv(glGetUniformLocation(m_lavaShader, "uNormalMatrix"), 1, GL_FALSE, value_ptr(normalMatrix));
-	glUniformMatrix4fv(glGetUniformLocation(m_lavaShader, "uViewMatrix"), 1, GL_FALSE, value_ptr(view));
-
-	// Pass inverse matrices for raymarching
-	mat4 invProj = inverse(proj);
-	mat4 invView = inverse(view);
-	glUniformMatrix4fv(glGetUniformLocation(m_lavaShader, "uInvProjectionMatrix"), 1, GL_FALSE, value_ptr(invProj));
-	glUniformMatrix4fv(glGetUniformLocation(m_lavaShader, "uInvViewMatrix"), 1, GL_FALSE, value_ptr(invView));
-
-	// Set time uniform
-	glUniform1f(glGetUniformLocation(m_lavaShader, "uTime"), static_cast<float>(glfwGetTime()));
-
-	// Set camera position (world-space)
-	vec3 cameraPos = vec3(inverse(view) * vec4(0, 0, 0, 1));
-	glUniform3fv(glGetUniformLocation(m_lavaShader, "uCameraPos"), 1, value_ptr(cameraPos));
-
-	// Pass framebuffer resolution
-	m_windowsize = vec2(width, height);
-	glUniform2fv(glGetUniformLocation(m_lavaShader, "uResolution"), 1, value_ptr(m_windowsize));
-
-	// Lamp parameters (use simulation getters so geometry + sim match)
-	glUniform1f(glGetUniformLocation(m_lavaShader, "uLampRadius"), m_lavaLamp.getRadius());
-	glUniform1f(glGetUniformLocation(m_lavaShader, "uLampTopRadius"), 1.0f); // adjust/getter if needed
-	glUniform1f(glGetUniformLocation(m_lavaShader, "uLampHeight"), m_lavaLamp.getHeight());
-	glUniform1f(glGetUniformLocation(m_lavaShader, "uThreshold"), m_threshold);
-
-	// Radius padding (small safety margin)
-	GLint locPad = glGetUniformLocation(m_lavaShader, "uRadiusPadding");
-	if (locPad != -1) {
-		float padding = glm::max(0.02f, 0.02f * m_lavaLamp.getRadius());
-		glUniform1f(locPad, padding + 0.02f);
-	}
-
-	// Lighting
-	vec3 lightPos = vec3(5.0f, 15.0f, 5.0f);
-	vec3 lightColor = vec3(1.0f, 1.0f, 1.0f);
-	vec3 ambientColor = vec3(0.2f, 0.1f, 0.1f);
-	glUniform3fv(glGetUniformLocation(m_lavaShader, "uLightPos"), 1, value_ptr(lightPos));
-	glUniform3fv(glGetUniformLocation(m_lavaShader, "uLightColor"), 1, value_ptr(lightColor));
-	glUniform3fv(glGetUniformLocation(m_lavaShader, "uAmbientColor"), 1, value_ptr(ambientColor));
-
-	// Blob data
-	auto positions = m_lavaLamp.getBlobPositions();
-	auto radii = m_lavaLamp.getBlobRadii();
-	auto blobbiness = m_lavaLamp.getBlobBlobbiness();
-	auto colors = m_lavaLamp.getBlobColors();
-	int blobCount = m_lavaLamp.getBlobCount();
-
-	glUniform1i(glGetUniformLocation(m_lavaShader, "uBlobCount"), blobCount);
-	if (blobCount > 0) {
-		int count = std::min(blobCount, 16);
-		glUniform4fv(glGetUniformLocation(m_lavaShader, "uBlobPositions"), count, value_ptr(positions[0]));
-		glUniform1fv(glGetUniformLocation(m_lavaShader, "uBlobRadii"), count, radii.data());
-		glUniform1fv(glGetUniformLocation(m_lavaShader, "uBlobBlobbiness"), count, blobbiness.data());
-		glUniform3fv(glGetUniformLocation(m_lavaShader, "uBlobColors"), count, value_ptr(colors[0]));
-	}
-
-	// PASS 1: Depth-only pre-pass (metal + glass write depth)
-	glEnable(GL_DEPTH_TEST);
-	glDepthFunc(GL_LESS);
-	glDepthMask(GL_TRUE);
-	glColorMask(GL_FALSE, GL_FALSE, GL_FALSE, GL_FALSE);
-
-	// Render opaque metal (writes depth)
-	glUniform1i(glGetUniformLocation(m_lavaShader, "uRenderMode"), 2);
-	m_lampMetalModel.draw(view, proj);
-
-
-	// PASS 2: Metaball raymarching
-	glColorMask(GL_TRUE, GL_TRUE, GL_TRUE, GL_TRUE);
-	glDepthFunc(GL_LESS); // Pass if metaball <= existing depth
-	glDepthMask(GL_TRUE);
-
-	glUniform1i(glGetUniformLocation(m_lavaShader, "uRenderMode"), 1);
-	glUniform1i(glGetUniformLocation(m_lavaShader, "uIsFullscreenQuad"), 1);
-
-	// Bind depth texture from metal pre-pass (read-only)
-	glActiveTexture(GL_TEXTURE0);
-	glBindTexture(GL_TEXTURE_2D, m_depthTextureFront);
-	glUniform1i(glGetUniformLocation(m_lavaShader, "uDepthTexture"), 0);
-
-	m_fullscreenQuadModel.draw(view, proj);
-	glUniform1i(glGetUniformLocation(m_lavaShader, "uIsFullscreenQuad"), 0);
-
-
-	// PASS 3: Metal color pass
-	glDepthFunc(GL_LEQUAL);
-	glDepthMask(GL_FALSE);
-
-	glUniform1i(glGetUniformLocation(m_lavaShader, "uRenderMode"), 2);
-	m_lampMetalModel.draw(view, proj);
-
-
-	// PASS 4: Glass color pass
-	glEnable(GL_BLEND);
-	glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
-	glDepthMask(GL_FALSE);
-	glDepthFunc(GL_LEQUAL);
-
-	glUniform1i(glGetUniformLocation(m_lavaShader, "uRenderMode"), 0);
-	m_lampGlassModel.draw(view, proj);
-
-	// Restore state
-	glDepthMask(GL_TRUE);
-	glDepthFunc(GL_LESS);
-	glDisable(GL_BLEND);
-
-	// RESTORE ALL STATE at the end:
-	glDepthMask(depthMask);
-	glDepthFunc(depthFunc);
-	glColorMask(GL_TRUE, GL_TRUE, GL_TRUE, GL_TRUE);
-
-	if (!blendEnabled) {
-		glDisable(GL_BLEND);
-	}
-
-	// IMPORTANT: Unbind textures and reset to default shader
-	glActiveTexture(GL_TEXTURE0);
-	glBindTexture(GL_TEXTURE_2D, 0);
-	glBindTexture(GL_TEXTURE_CUBE_MAP, 0);
-	glUseProgram(0);
-
-
-}
 
 void Application::render() {
 	// retrieve the window hieght
@@ -671,7 +126,11 @@ void Application::render() {
 	glPolygonMode(GL_FRONT_AND_BACK, (m_showWireframe) ? GL_LINE : GL_FILL);
 
 	// Render lava lamp
-	renderLavaLamp(view, proj);
+	m_lavaLamp.renderLavaLamp(
+		view, proj, m_window,
+		m_animateLamp, m_showLavaLamp, m_threshold,
+		m_heaterTemp, m_gravity
+	);
 
 
 	// ----------------------------
